Fix Time Series Merge
This commit repairs a bug in the C++ engine merge logic ; previously, it was not
properly sorting and accumulating the samples from the first merge operation on
a given key, instead just accepting the collection exactly as it was provided.

This commit also exports the C++ merging logic for Time Series as the
"MergeInternalTimeSeriesData" method in the engine package. This method is being
exported for usage in higher-level time series tests which are forthcoming.